Share cooldown_waste_data_t class
Move mage implementation into its own file, integrate into player_t.

Remove paladin cooldown waste factory.

No changes to report output yet.
