+++ engine/class_modules/paladin/sc_paladin.hpp
@@ -50,84 +50,9 @@ struct paladin_td_t : public actor_target_data_t
   paladin_td_t( player_t* target, paladin_t* paladin );
 };
 
-struct cooldown_waste_data_t : public noncopyable
-{
-  const cooldown_t* cd;
-  double buffer;
-
-  extended_sample_data_t normal;
-  extended_sample_data_t cumulative;
-
-  cooldown_waste_data_t( const cooldown_t* cooldown, bool simple = true ) :
-    cd( cooldown ), buffer( 0.0 ), normal( cd -> name_str + " waste", simple ),
-    cumulative( cd -> name_str + " cumulative waste", simple ) {}
-
-  virtual bool may_add( timespan_t cd_override = timespan_t::min() ) const
-  {
-    return ( cd -> duration > 0_ms || cd_override > 0_ms )
-        && ( ( cd -> charges == 1 && cd -> up() ) || ( cd -> charges >= 2 && cd -> current_charge == cd -> charges ) )
-        && ( cd -> last_charged > 0_ms && cd -> last_charged < cd -> sim.current_time() );
-  }
-
-  virtual double get_wasted_time()
-  {
-    return (cd -> sim.current_time() - cd -> last_charged).total_seconds();
-  }
-
-  void add( timespan_t cd_override = timespan_t::min(), timespan_t time_to_execute = 0_ms )
-  {
-    if ( may_add( cd_override ) )
-    {
-      double wasted = get_wasted_time();
-      if ( cd -> charges == 1 )
-      {
-        wasted -= time_to_execute.total_seconds();
-      }
-      normal.add( wasted );
-      buffer += wasted;
-    }
-  }
-
-  bool active() const
-  {
-    return normal.count() > 0 && cumulative.sum() > 0;
-  }
-
-  void merge( const cooldown_waste_data_t& other )
-  {
-    normal.merge( other.normal );
-    cumulative.merge( other.cumulative );
-  }
-
-  void analyze()
-  {
-    normal.analyze();
-    cumulative.analyze();
-  }
-
-  void datacollection_begin()
-  {
-    buffer = 0.0;
-  }
-
-  void datacollection_end()
-  {
-    if ( may_add() )
-      buffer += get_wasted_time();
-    cumulative.add( buffer );
-    buffer = 0.0;
-  }
-
-  virtual ~cooldown_waste_data_t() { }
-};
-
 struct paladin_t : public player_t
 {
 public:
-
-  // waste tracking
-  auto_dispose<std::vector<cooldown_waste_data_t*> > cooldown_waste_data_list;
-
   // Active spells
   struct active_spells_t
   {
@@ -626,28 +551,6 @@ public:
 
   virtual const paladin_td_t* find_target_data( const player_t* target ) const override;
   virtual paladin_td_t* get_target_data( player_t* target ) const override;
-
-  cooldown_waste_data_t* get_cooldown_waste_data( cooldown_t* cd, cooldown_waste_data_t *(*factory)(cooldown_t*) = nullptr )
-  {
-    for ( auto cdw : cooldown_waste_data_list )
-    {
-      if ( cdw -> cd -> name_str == cd -> name_str )
-        return cdw;
-    }
-
-    cooldown_waste_data_t* cdw = nullptr;
-    if ( factory == nullptr ) {
-      cdw = new cooldown_waste_data_t( cd );
-    } else {
-      cdw = factory( cd );
-    }
-    cooldown_waste_data_list.push_back( cdw );
-    return cdw;
-  }
-  virtual void merge( player_t& other ) override;
-  virtual void analyze( sim_t& s ) override;
-  virtual void datacollection_begin() override;
-  virtual void datacollection_end() override;
 };
 
 namespace buffs {
@@ -766,7 +669,6 @@ public:
 
   bool track_cd_waste;
   cooldown_waste_data_t* cd_waste;
-  cooldown_waste_data_t* (*cd_waste_factory)(cooldown_t *);
 
   // Damage increase whitelists
   struct affected_by_t
@@ -784,7 +686,7 @@ public:
                     const spell_data_t* s = spell_data_t::nil() ) :
     ab( n, p, s ),
     track_cd_waste( s -> cooldown() > 0_ms || s -> charge_cooldown() > 0_ms ),
-    cd_waste( nullptr ), cd_waste_factory( nullptr ),
+    cd_waste( nullptr ), 
     affected_by( affected_by_t() ),
     hasted_cd( false ), hasted_gcd( false )
   {
@@ -826,7 +728,7 @@ public:
 
     if ( track_cd_waste && ab::sim -> report_details != 0 )
     {
-      cd_waste = p() -> get_cooldown_waste_data( ab::cooldown, cd_waste_factory );
+      cd_waste = p() -> get_cooldown_waste_data( ab::cooldown );
     }
 
     if ( hasted_cd )
