+++ source_files/QT_engine.pri
@@ -135,6 +135,7 @@ HEADERS += engine/report/reports.hpp
 HEADERS += engine/report/sc_highchart.hpp
 HEADERS += engine/sc_enums.hpp
 HEADERS += engine/sim/benefit.hpp
+HEADERS += engine/sim/cooldown_waste_data.hpp
 HEADERS += engine/sim/event.hpp
 HEADERS += engine/sim/event_manager.hpp
 HEADERS += engine/sim/gain.hpp
@@ -309,6 +310,7 @@ SOURCES += engine/report/sc_highchart.cpp
 SOURCES += engine/report/sc_report_html_player.cpp
 SOURCES += engine/report/sc_report_html_sim.cpp
 SOURCES += engine/report/sc_report_text.cpp
+SOURCES += engine/sim/cooldown_waste_data.cpp
 SOURCES += engine/sim/event_manager.cpp
 SOURCES += engine/sim/proc.cpp
 SOURCES += engine/sim/real_ppm.cpp
